Moves Room-related classes to Rooms namespace.
