+++ ParquetClassLibrary/Rooms/Space.cs
@@ -1,7 +1,7 @@
 ﻿using ParquetClassLibrary.Parquets;
 using ParquetClassLibrary.Stubs;
 
-namespace ParquetClassLibrary.Map
+namespace ParquetClassLibrary.Rooms
 {
     /// <summary>
     /// A <see cref="ParquetStack"/> together with its coordinates on the world map.
